update bicep, update azure ai variable names, update gpt4o in speech samples, update speech functions, copy nl to sql sample
